pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tldr-crypto-finance"
version = "0.1.0"
description = "Local-first finance and newsletter intelligence pipeline."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12",
  "duckdb>=1.0",
  "pydantic-settings>=2.2",
  "PyYAML>=6.0",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "ipykernel>=6.29",
  "pytest>=8.0",
  "ruff>=0.6",
]
gmail = [
  "google-api-python-client>=2.140",
  "google-auth-httplib2>=0.2",
  "google-auth-oauthlib>=1.2",
]
ml = [
  "sentence-transformers>=3.0",
  "transformers>=4.44",
]

[project.scripts]
tldr-crypto-finance = "tldr_crypto_finance.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
